% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_tools.R
\name{gis_tools}
\alias{gis_tools}
\alias{gh_to_sf}
\alias{gh_to_sf.default}
\alias{gh_to_sf.data.frame}
\alias{gh_covering}
\title{Helpers for interfacing geohashes with sf objects}
\usage{
gh_to_sf(...)

\method{gh_to_sf}{default}(geohashes, ...)

\method{gh_to_sf}{data.frame}(gh_df, gh_col = "gh", ...)

gh_covering(x, precision = 6L, minimal = FALSE)
}
\arguments{
\item{\dots}{Arguments for subsequent methods.}

\item{geohashes}{\code{character} vector of geohashes to be converted to polygons.}

\item{gh_df}{\code{data.frame} which 1) contains a column of geohashes to be converted to polygons and 2) will
serve as the attribute table of the resultant \link[sf:sf]{sf::sf} object.}

\item{gh_col}{\code{character} column name saying where the geohashes are stored in \code{gh_df}.}

\item{x}{An \code{sf}, \code{sfc}, or \code{sfg} object (from the \code{sf} package).}

\item{precision}{\code{integer} specifying the precision of geohashes to use, same as \code{\link[=gh_encode]{gh_encode()}}}

\item{minimal}{\code{logical}; if \code{FALSE}, the output will have all geohashes in the bounding box of \code{x}; if
\code{TRUE}, any geohashes not intersecting \code{x} will be removed.}
}
\value{
For \code{gh_to_sf}, a \link[sf:sf]{sf::sf} object.
}
\description{
These functions smooth the gateway between working with geohashes and geospatial information built for the
major geospatial package in R, \link[sf:sf]{sf::sf}.
}
\details{
\code{gh_to_sf} relies on the \code{\link[=gh_decode]{gh_decode()}} function. Note in particular that this function accepts any length
of geohash (geohash-6, geohash-4, etc.) and is agnostic to potential overlap, though duplicates will be caught
and excluded.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# get the neighborhood of this geohash in downtown Apia as an sf object
downtown = '2jtc5x'
apia_nbhd = unlist(gh_neighbors(downtown))
apia_sf = gh_to_sf(apia_nbhd)

# all geohashes covering a random sampling within Apia:
apia_covering = gh_covering(sf::st_sample(apia_sf, 10L, type = 'random'))
\dontshow{\}) # examplesIf}
}
